FOUR-90 Feat : Follow Session 검사 코드 추가

+++ src/main/java/com/hallym/booker/controller/FollowApiController.java
@@ -1,10 +1,14 @@
 package com.hallym.booker.controller;
 
 import com.hallym.booker.domain.Profile;
+import com.hallym.booker.domain.SessionConst;
 import com.hallym.booker.dto.Follow.FollowRequest;
 import com.hallym.booker.dto.Follow.LatestJournalsResponse;
 import com.hallym.booker.dto.Follow.ProfileDto;
+import com.hallym.booker.dto.Login.LoginResponse;
 import com.hallym.booker.service.FollowService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,17 +27,35 @@ public class FollowApiController {
 
     private final FollowService followService;
 
+
     // 팔로잉 & 팔로워 추가
     @PostMapping("/new")
-    public ResponseEntity<String> addFollow(@RequestBody @Valid final FollowRequest followRequest) {
+    public ResponseEntity<String> addFollow(@RequestBody @Valid final FollowRequest followRequest, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if(session == null){ //세션이 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        LoginResponse loginResponse = (LoginResponse) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        if(loginResponse == null){ //세션에 회원 데이터가 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
         followService.newFollow(followRequest.getFromUserId(), followRequest.getToUserId());
         return new ResponseEntity<>("Follow created successfully", HttpStatus.OK);
     }
 
+
     // 팔로잉 목록 조회 - 프로필(사진 이름, URL)&닉네임 (전체)
     @PostMapping("/followingsList")
-    public ResponseEntity<List<ProfileDto>> followingsList(@RequestBody final Long fromUserId) {
-        List<Profile> followings = followService.findAllToUserIdProfile(fromUserId);
+    public ResponseEntity<List<ProfileDto>> followingsList(HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if(session == null){ //세션이 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        LoginResponse loginResponse = (LoginResponse) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        if(loginResponse == null){ //세션에 회원 데이터가 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        List<Profile> followings = followService.findAllToUserIdProfile(loginResponse.getUid());
         List<ProfileDto> followingsInfo = new ArrayList<>();
         for(Profile profile : followings){
             followingsInfo.add(new ProfileDto(profile.getProfileUid(),profile.getNickname(),
@@ -44,8 +66,16 @@ public class FollowApiController {
 
     // 팔로워 목록 조회 - 프로필(사진 이름, URL)&닉네임 (전체)
     @PostMapping("/followersList")
-    public ResponseEntity<List<ProfileDto>> followersList(@RequestBody final Long toUserId) {
-        List<Profile> followers = followService.findAllFromUserIdProfile(toUserId);
+    public ResponseEntity<List<ProfileDto>> followersList(HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if(session == null){ //세션이 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        LoginResponse loginResponse = (LoginResponse) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        if(loginResponse == null){ //세션에 회원 데이터가 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        List<Profile> followers = followService.findAllFromUserIdProfile(loginResponse.getUid());
         List<ProfileDto> followersInfo = new ArrayList<>();
         for(Profile profile : followers){
             followersInfo.add(new ProfileDto(profile.getProfileUid(),profile.getNickname(),
@@ -56,20 +86,44 @@ public class FollowApiController {
 
     // 팔로잉들의 최신 독서록 목록 조회 - 프로필(사진 이름, URL)&닉네임
     @PostMapping("/followingsLatestJournals")
-    public ResponseEntity<List<LatestJournalsResponse>> followingsLatestJournals(@RequestBody final Long fromUserId) {
-        return ResponseEntity.ok().body(followService.findFollowingsLatestJournals(fromUserId));
+    public ResponseEntity<List<LatestJournalsResponse>> followingsLatestJournals(HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if(session == null){ //세션이 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        LoginResponse loginResponse = (LoginResponse) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        if(loginResponse == null){ //세션에 회원 데이터가 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        return ResponseEntity.ok().body(followService.findFollowingsLatestJournals(loginResponse.getUid()));
     }
 
     // 팔로잉 삭제
     @PostMapping("/delete")
-    public ResponseEntity<String> deleteFollow(@RequestBody @Valid final FollowRequest followRequest) {
+    public ResponseEntity<String> deleteFollow(@RequestBody @Valid final FollowRequest followRequest, HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if(session == null){ //세션이 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        LoginResponse loginResponse = (LoginResponse) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        if(loginResponse == null){ //세션에 회원 데이터가 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
         followService.removeFollowing(followRequest.getFromUserId(),followRequest.getToUserId());
         return new ResponseEntity<>("Follow deleted successfully",HttpStatus.OK);
     }
 
     //팔로잉 유무 확인
     @PostMapping("/followCheck")
-    public ResponseEntity<Boolean> followCheck(@RequestBody @Valid final FollowRequest followRequest){
+    public ResponseEntity<Boolean> followCheck(@RequestBody @Valid final FollowRequest followRequest, HttpServletRequest request){
+        HttpSession session = request.getSession(false);
+        if(session == null){ //세션이 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
+        LoginResponse loginResponse = (LoginResponse) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        if(loginResponse == null){ //세션에 회원 데이터가 없으면 홈으로 이동하게 null
+            return new ResponseEntity<>(null, HttpStatus.FOUND);
+        }
         return ResponseEntity.ok().body(followService.checkFollowing(followRequest.getFromUserId(),followRequest.getToUserId()));
     }
 }
